requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
pgvector>=0.2.0
pydantic>=2.0.0
tree-sitter>=0.20.1
numpy>=1.24.0
httpx>=0.24.0
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
requests>=2.31.0